backend/requirements.txt
Django==3.1.4
django-3-jet==1.0.8
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.6.0
django-import-export==2.5.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==8.0.1
psycopg2==2.8.6
stripe==2.56.0
